Move images from public to images
