+++ packages/arb-token-bridge-ui/src/components/TransferPanel/TokenImportDialog.tsx
@@ -1,25 +1,28 @@
-import { useState, useEffect, useMemo, useCallback } from 'react'
-import { useLatest } from 'react-use'
-import { ERC20BridgeToken, getL1TokenData } from 'token-bridge-sdk'
 import {
   ExclamationCircleIcon,
   InformationCircleIcon
 } from '@heroicons/react/outline'
 import Tippy from '@tippyjs/react'
-import { Loader } from '../common/atoms/Loader'
-import { useActions, useAppState } from '../../state'
-import {
-  useTokensFromLists,
-  useTokensFromUser,
-  toERC20BridgeToken
-} from './TokenSearchUtils'
-import { useNetworksAndSigners } from '../../hooks/useNetworksAndSigners'
+
+import Image from 'next/image'
+import { useCallback, useEffect, useMemo, useState } from 'react'
+import { useLatest } from 'react-use'
+import { ERC20BridgeToken, getL1TokenData } from 'token-bridge-sdk'
+
 import { useERC20L1Address } from '../../hooks/useERC20L1Address'
+import { useNetworksAndSigners } from '../../hooks/useNetworksAndSigners'
+import { useActions, useAppState } from '../../state'
+import { getExplorerUrl } from '../../util/networks'
+import { Loader } from '../common/atoms/Loader'
 import { Dialog, UseDialogProps } from '../common/Dialog'
 import { SafeImage } from '../common/SafeImage'
-import { getExplorerUrl } from '../../util/networks'
-import Image from 'next/image'
 import GrumpyCat from '@/images/grumpy-cat.webp'
+
+import {
+  toERC20BridgeToken,
+  useTokensFromLists,
+  useTokensFromUser
+} from './TokenSearchUtils'
 
 enum ImportStatus {
   LOADING,
