+++ packages/arb-token-bridge-ui/src/components/TransferPanel/WithdrawalConfirmationDialog.tsx
@@ -24,7 +24,7 @@ import {
 } from '../../util/networks'
 import { getFastBridges } from '../../util/fastBridges'
 import Image from 'next/image'
-import GoogleCalendarSVG from '../../../public/GoogleCalendar.svg'
+import GoogleCalendarSVG from '@/images/GoogleCalendar.svg'
 
 const SECONDS_IN_DAY = 86400
 const SECONDS_IN_HOUR = 3600
